% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouse.R
\docType{data}
\name{raw_data_factors_cols}
\alias{raw_data_factors_cols}
\title{Local Job Id',}
\format{
An object of class \code{list} of length 3.
}
\usage{
raw_data_factors_cols
}
\description{
Local Job Id',
}
\keyword{datasets}
