% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxdmod-package.R
\docType{package}
\name{rxdmod-package}
\alias{rxdmod}
\alias{rxdmod-package}
\title{rxdmod: XDMoD Data Analytics Framework}
\description{
As part of the XDMoD Data Analytics Framework, this R package provides API access to the data warehouse of an instance of XDMoD version ≥10.5.
}
\author{
\strong{Maintainer}: Nikolay Simakov \email{nikolays@buffalo.edu} (\href{https://orcid.org/0000-0001-6131-5979}{ORCID})

Authors:
\itemize{
  \item Aaron Weeden \email{aaronwee@buffalo.edu}
  \item Joseph White \email{jpwhite4@buffalo.edu}
}

}
\keyword{internal}
