% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_xdmod}
\alias{theme_xdmod}
\title{Set XDMoD-like theme}
\usage{
theme_xdmod(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  set_colors = T
)
}
\arguments{
\item{set_colors}{set default colors to XDMoD pallet}
}
\value{
ggplot2 theme
}
\description{
Set XDMoD-like theme, which is just \code{ggplot2::theme_bw} at this point
}
\examples{
theme_set(theme_xdmod(set_colors = T))

dw <- xdmod_get_datawarehouse('https://xdmod.access-ci.org')

df <- xdmod_get_data(dw,
  duration=c('2023-01-01', '2023-03-31'),
  realm='Jobs',
  metric='CPU Hours: Total',
  dimension='Resource'
)

ggplot(df,aes(x=Time, y=`CPU Hours: Total`, color=Resource)) + geom_line()

}
