% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouse.R
\name{xdmod_describe_dimensions}
\alias{xdmod_describe_dimensions}
\title{Title}
\usage{
xdmod_describe_dimensions(datawarehouse, realm)

xdmod_describe_dimensions(datawarehouse, realm)
}
\arguments{
\item{realm}{}
}
\description{
Title

Title
}
