% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouse.R
\name{xdmod_get_datawarehouse}
\alias{xdmod_get_datawarehouse}
\title{Get XDMoD data-warehouse connector}
\usage{
xdmod_get_datawarehouse(xdmod_host)
}
\arguments{
\item{xdmod_host}{URL to XDMoD instance, for example 'https://xdmod.access-ci.org'}
}
\value{
XDMoD connector
}
\description{
Get XDMoD data-warehouse connector.
Enviroment variable \code{XDMOD_API_TOKEN} should contain XDMoD API token.
}
\details{
Get the token from your XDMoD instance and store in ~/.Renviron (~/Documents/.Renviron on windows) as:

\verb{XDMOD_API_TOKEN=<secret key>}
}
\examples{
dw <- xdmod_get_datawarehouse('https://xdmod.access-ci.org')
df <- xdmod_get_data(dw,
    duration=c('2023-01-01', '2023-04-30'),
    realm='Jobs',
    metric='Number of Users: Active'
    )

}
