% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouse.R
\name{xdmod_get_filter_values}
\alias{xdmod_get_filter_values}
\title{Title}
\usage{
xdmod_get_filter_values(datawarehouse, realm, dimension)

xdmod_get_filter_values(datawarehouse, realm, dimension)
}
\arguments{
\item{dimension}{}
}
\description{
Title

Title
}
