% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouse.R
\name{xdmod_get_raw_data}
\alias{xdmod_get_raw_data}
\title{Title}
\usage{
xdmod_get_raw_data(
  datawarehouse,
  duration,
  realm,
  fields = NULL,
  filters = list(),
  show_progress = FALSE,
  convert_timestamp = TRUE,
  tz = NULL,
  strings_as_factors = FALSE
)
}
\arguments{
\item{strings_as_factors}{}
}
\description{
Title
}
