% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{xdmod_pal}
\alias{xdmod_pal}
\title{XDMoD color palette (discrete)}
\usage{
xdmod_pal()
}
\arguments{
\item{palette}{\code{character} The name of the xdmod theme to use. One of
\code{"default"}, or \code{"darkunica"}.}
}
\description{
The Highcharts uses many different color palettes in its
plots. This collects a few of them.
}
\seealso{
Other colour xdmod: 
\code{\link{scale_colour_xdmod}()}
}
\concept{colour xdmod}
