% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{xdmod_set_colors}
\alias{xdmod_set_colors}
\title{Set default option to use XDMoD pallet}
\usage{
xdmod_set_colors()
}
\description{
Set default option to use XDMoD pallet
}
\examples{
xdmod_set_colors()

dw <- xdmod_get_datawarehouse('https://xdmod.access-ci.org')

df <- xdmod_get_data(dw,
  duration=c('2023-01-01', '2023-03-31'),
  realm='Jobs',
  metric='CPU Hours: Total',
  dimension='Resource'
)

ggplot(df,aes(x=Time, y=`CPU Hours: Total`, color=Resource)) + geom_line()


}
